+++ CHANGES.md
@@ -20,6 +20,10 @@ This file is a best-effort approach to solving this issue; we will do our best b
 * (wifi) Changes have been made to the `WifiRemoteStationManager` interface for what concerns the update of the frame retry count of the MPDUs and the decision of dropping MPDUs (possibly based on the max retry limit). The `NeedRetransmission` method has been replaced by the `GetMpdusToDropOnTxFailure` method and the `DoNeedRetransmission` method has been replaced by the `DoGetMpdusToDropOnTxFailure` method. Also, the `DoIncrementRetryCountOnTxFailure` method has been added to implement custom policies for the update of the frame retry count of MPDUs upon transmission failure.
 * (applications) Added an `OnOffState` trace source to `OnOffApplication`, to track whether the application is transmitting or not.
 * (zigbee) Added Zigbee module support. The module includes a NWK layer with joining and routing capabilities. No APS layer included.
+* (mobility) Added `MobilityModel::GetWraparoundModel()`, `MobilityModel::SetWraparoundModel()` and `MobilityModelHelper::SetWraparoundModel()`, which configure a wraparound model to the mobility model, or retrieves it.
+* (mobility) Added `MobilityModel::GetVirtualPosition(const Vector& ref)` function to calculate the virtual position of a node relative to a coordinate `ref`, using the configured wraparound model. In case no wraparound is configured, `MobilityModel::DoGetPosition()` is returned instead.
+* (mobility) Added a virtual base class `WraparoundModel` for wraparound models.
+* (mobility) Added the `HexagonalWraparoundModel` class with the wraparound implementation for hexagonal cell deployments with 0, 1 or 3 rings.
 
 ### Changes to existing API